feat: update fsxa-api and implement new fsxa-api classes (#17)
The new version of the fsxa-api (v6.0.0-alpha.1) uses new classes for the fsxa-api. These were
implemented. Furthermore the store was enhanced, so that the navigation and an auth object can be
changed.

BREAKING CHANGE: The original fsxa-api class was removed and the new ones FSXAProxyApi and
FSXARemoteApi were implemented. Their have a slightly different, but better, method signatures. For
more information, please read the README file.