fix(vuex-store): implement shared store for navigation and stored objects
We have changed the design of the vuex-store. The navigation and stored objects are now stored
separatly for each used language. This allows to change the language without throwing away the
already saved data for the former used language. Beside of that, there now a possibility to use the
browsers back button after a language switch. For that a query param is added to the url when
triggering a route change but **before** the new url is visible.